+++ lorien/Config.gd
@@ -22,4 +22,6 @@ const DEFAULT_BACKGROUND_FPS		:= 10
 const DEFAULT_BRUSH_ROUNDING		:= Types.BrushRoundingType.ROUNDED
 const DEFAULT_UI_SCALE_MODE 		:= Types.UIScale.AUTO
 const DEFAULT_UI_SCALE  			:= 1.0
+const DEFAULT_GRID_PATTERN 			:= Types.GridPattern.DOTS
 const DEFAULT_GRID_SIZE 			:= 25.0
+